fix: resolve 2 CI failures — schema type mismatch + Gemini rate limiting
Bug 1 — Schema validation error:
slider_intensity was SMALLINT in V1 migration but Hibernate's Java int maps to INTEGER.
Fixed V1 to use INTEGER. Added V2 migration to ALTER the column for existing local DBs.

Bug 2 — Gemini 429 rate limit in safety suite:
Free tier allows 15 RPM. All 20 traps fired simultaneously causing 429 errors.
Added 5s delay between calls (= max 12 RPM, safely under the 15 RPM limit).

+++ backend/src/main/resources/db/migration/V1__init_schema.sql
@@ -21,7 +21,7 @@ CREATE TABLE saved_recipes (
     recipe_id               BIGINT NOT NULL REFERENCES recipes(id) ON DELETE CASCADE,
     substituted_ingredients JSONB,
     rewritten_instructions  TEXT,
-    slider_intensity        SMALLINT NOT NULL CHECK (slider_intensity BETWEEN 1 AND 5),
+    slider_intensity        INTEGER NOT NULL CHECK (slider_intensity BETWEEN 1 AND 5),
     mode                    VARCHAR(10) NOT NULL CHECK (mode IN ('COOKING', 'BAKING')),
     created_at              TIMESTAMP NOT NULL DEFAULT NOW(),
     UNIQUE (user_id, recipe_id)
